Refactored main.py.
